added screen capturer sample

+++ sample-screencapturer/src/main/java/com/quickblox/screencapturer/fragments/ScreenCaptiringFragment.java
@@ -73,7 +73,7 @@ public class ScreenCaptiringFragment extends Fragment implements View.OnClickLis
     }
 
     @Override
-    public void OnFrameCaptured(int[] data, int width, int height, int rotation, long timeStamp) {
+    public void OnFrameCaptured(byte[] data, int width, int height, int rotation, long timeStamp) {
 
     }
